LayoutTests/http/tests/security/contentSecurityPolicy/object-src-no-url-blocked.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/plugin.js"></script>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();

    runAfterPluginLoad(null, NotifyDone);
</script>
<meta http-equiv="Content-Security-Policy" content="object-src 'none'">
</head>
<body>
This test passes if there is a console message saying the plugin was blocked.
<object type="application/x-webkit-test-netscape"></object>
</body>
</html>
