LayoutTests/http/tests/security/contentSecurityPolicy/object-src-no-url-blocked-expected.txt

CONSOLE MESSAGE: line 1: Refused to load object from '' because of Content-Security-Policy.

This test passes if there is a console message saying the plugin was blocked. 
